chore: move influxql e2e test to cloud
